Refactor table utils
