+++ src/components/SlateEditor/plugins/table/index.tsx
@@ -20,13 +20,6 @@ import {
   removeEmptyElementDataAttributes,
 } from '../../../../util/embedTagHelpers';
 import SlateTable from './SlateTable';
-import {
-  defaultTableBodyBlock,
-  defaultTableCaptionBlock,
-  defaultTableCellBlock,
-  defaultTableHeadBlock,
-  defaultTableRowBlock,
-} from './utils';
 import getCurrentBlock from '../../utils/getCurrentBlock';
 import { normalizeTableBodyAsMatrix } from './matrix';
 import { handleTableKeydown } from './handleKeyDown';
@@ -62,6 +55,13 @@ import {
   isTableRow,
 } from './slateHelpers';
 import { updateCell } from './slateActions';
+import {
+  defaultTableBodyBlock,
+  defaultTableCaptionBlock,
+  defaultTableCellBlock,
+  defaultTableHeadBlock,
+  defaultTableRowBlock,
+} from './defaultBlocks';
 
 const validKeys = [KEY_ARROW_UP, KEY_ARROW_DOWN, KEY_TAB, KEY_BACKSPACE, KEY_DELETE];
